Because there are no indices for full-text search, searching for a job ID takes an inordinate amount of time. This commit examines the search query and if the only item is a UUID, it specifically searches for a job matching that ID.

Fixes #1618