@francois 2025-03-12
This is tested and generates this SQL query:

```sql
SELECT "good_jobs"."id", "good_jobs"."queue_name", "good_jobs"."priority", "good_jobs"."serialized_params", "good_jobs"."scheduled_at", "good_jobs"."performed_at", "good_jobs"."finished_at", "good_jobs"."error", "good_jobs"."created_at", "good_jobs"."updated_at", "good_jobs"."active_job_id", "good_jobs"."concurrency_key", "good_jobs"."cron_key", "good_jobs"."cron_at", "good_jobs"."batch_id", "good_jobs"."batch_callback_id", "good_jobs"."executions_count", "good_jobs"."job_class", "good_jobs"."error_event", "good_jobs"."labels", "good_jobs"."locked_by_id", "good_jobs"."locked_at" FROM "good_jobs" WHERE "good_jobs"."active_job_id" = '38979b0b-af58-4bf4-969e-017b10fa3bb7' ORDER BY created_at desc, id desc LIMIT 25
```

We can argue that `active_job_id` is not the right field to search for, but that's an easy fix.

Another thing I was thinking of: should we search for the execution ID at the same time? When I'm querying by job ID, it's usually because I grabbed a job ID from PaperTrail and I'm looking for more details.