@francois 2025-03-15
Hey @bensheldon, I made some research on this fine Saturday morning. Here's a dump of my notes from the research. Feel free to use this research or not!

# Improving GoodJobs search

## Initial State

I copied our production database to development and gathered some statistics. Before gathering these statistics, I `VACUUM FULL ANALYZE` the `good_jobs` table. We are running GoodJobs 4.9.3, and we are fully migrated. We do not (yet?) use the labels feature.

```
# select version();
                                                          version
---------------------------------------------------------------------------------------------------------------------------
 PostgreSQL 17.4 (Debian 17.4-1.pgdg120+2) on aarch64-unknown-linux-gnu, compiled by gcc (Debian 12.2.0-14) 12.2.0, 64-bit

# select count(*) from good_jobs;
 count
--------
 174999

# select name, pg_size_pretty(pg_total_relation_size(name))
  from (values
    ('good_jobs')
  , ('good_jobs_pkey')
  , ('index_good_job_jobs_for_candidate_lookup')
  , ('index_good_jobs_jobs_on_finished_at')
  , ('index_good_jobs_jobs_on_priority_created_at_when_unfinished')
  , ('index_good_jobs_on_active_job_id_and_created_at')
  , ('index_good_jobs_on_batch_callback_id')
  , ('index_good_jobs_on_batch_id')
  , ('index_good_jobs_on_concurrency_key_and_created_at')
  , ('index_good_jobs_on_concurrency_key_when_unfinished')
  , ('index_good_jobs_on_cron_key_and_created_at_cond')
  , ('index_good_jobs_on_cron_key_and_cron_at_cond')
  , ('index_good_jobs_on_labels')
  , ('index_good_jobs_on_locked_by_id')
  , ('index_good_jobs_on_priority_scheduled_at_unfinished_unlocked')
  , ('index_good_jobs_on_queue_name_and_scheduled_at')
  , ('index_good_jobs_on_scheduled_at')) t0(name)
  ;

                             name                             | pg_size_pretty
--------------------------------------------------------------+----------------
 good_jobs                                                    | 142 MB
 good_jobs_pkey                                               | 5416 kB
 index_good_job_jobs_for_candidate_lookup                     | 48 kB
 index_good_jobs_jobs_on_finished_at                          | 3840 kB
 index_good_jobs_jobs_on_priority_created_at_when_unfinished  | 48 kB
 index_good_jobs_on_active_job_id_and_created_at              | 6960 kB
 index_good_jobs_on_batch_callback_id                         | 16 kB
 index_good_jobs_on_batch_id                                  | 200 kB
 index_good_jobs_on_concurrency_key_and_created_at            | 5472 kB
 index_good_jobs_on_concurrency_key_when_unfinished           | 16 kB
 index_good_jobs_on_cron_key_and_created_at_cond              | 224 kB
 index_good_jobs_on_cron_key_and_cron_at_cond                 | 224 kB
 index_good_jobs_on_labels                                    | 16 kB
 index_good_jobs_on_locked_by_id                              | 8192 bytes
 index_good_jobs_on_priority_scheduled_at_unfinished_unlocked | 40 kB
 index_good_jobs_on_queue_name_and_scheduled_at               | 40 kB
 index_good_jobs_on_scheduled_at                              | 32 kB
```

## GIN index exploration

From [Using JSON object values in PostgreSQL GIN index](https://stackoverflow.com/a/42693095), I copied and adapted `jsonb_values`, while renaming it to `good_jobs_jsonb_values`. Then I used that function to create a GIN index on `serialized_params` and `labels`.

```
CREATE OR REPLACE FUNCTION public.good_jobs_jsonb_values(jsonb)
RETURNS text[]
LANGUAGE plpgsql
IMMUTABLE
AS $$
DECLARE
    result text[];
    elem jsonb;
    val text;
BEGIN
    CASE jsonb_typeof($1)
        WHEN 'null' THEN
            RETURN ARRAY[]::text[];
        WHEN 'object' THEN
            FOR elem IN SELECT value FROM jsonb_each($1)
            LOOP
                result := result || public.good_jobs_jsonb_values(elem);
            END LOOP;
        WHEN 'array' THEN
            FOR elem IN SELECT jsonb_array_elements($1)
            LOOP
                result := result || public.good_jobs_jsonb_values(elem);
            END LOOP;
        ELSE
            -- Convert scalar values to text and return as array
            SELECT jsonb_array_elements_text(jsonb_build_array($1)) INTO val;
            RETURN ARRAY[val];
    END CASE;

    RETURN result;
END;
$$;

# CREATE INDEX index_good_jobs_on_serialized_params_and_labels ON good_jobs USING GIN ( (public.good_jobs_jsonb_values(serialized_params) || labels) );
CREATE INDEX
Time: 19171.844 ms (00:19.172)
```

From experience, I know that my MacBook Pro is about 2x faster than our production instance.

After creating the index, I checked the table sizes again:

```
# select name, pg_size_pretty(pg_total_relation_size(name))
  from (values
    ('good_jobs')
  , ('good_jobs_pkey')
  , ('index_good_job_jobs_for_candidate_lookup')
  , ('index_good_jobs_jobs_on_finished_at')
  , ('index_good_jobs_jobs_on_priority_created_at_when_unfinished')
  , ('index_good_jobs_on_active_job_id_and_created_at')
  , ('index_good_jobs_on_batch_callback_id')
  , ('index_good_jobs_on_batch_id')
  , ('index_good_jobs_on_concurrency_key_and_created_at')
  , ('index_good_jobs_on_concurrency_key_when_unfinished')
  , ('index_good_jobs_on_cron_key_and_created_at_cond')
  , ('index_good_jobs_on_cron_key_and_cron_at_cond')
  , ('index_good_jobs_on_labels')
  , ('index_good_jobs_on_locked_by_id')
  , ('index_good_jobs_on_priority_scheduled_at_unfinished_unlocked')
  , ('index_good_jobs_on_queue_name_and_scheduled_at')
  , ('index_good_jobs_on_scheduled_at')
  , ('index_good_jobs_on_serialized_params_and_labels')) t0(name)
  ;
                             name                             | pg_size_pretty
--------------------------------------------------------------+----------------
 good_jobs                                                    | 190 MB
 good_jobs_pkey                                               | 5416 kB
 index_good_job_jobs_for_candidate_lookup                     | 48 kB
 index_good_jobs_jobs_on_finished_at                          | 3840 kB
 index_good_jobs_jobs_on_priority_created_at_when_unfinished  | 48 kB
 index_good_jobs_on_active_job_id_and_created_at              | 6960 kB
 index_good_jobs_on_batch_callback_id                         | 16 kB
 index_good_jobs_on_batch_id                                  | 200 kB
 index_good_jobs_on_concurrency_key_and_created_at            | 5472 kB
 index_good_jobs_on_concurrency_key_when_unfinished           | 16 kB
 index_good_jobs_on_cron_key_and_created_at_cond              | 224 kB
 index_good_jobs_on_cron_key_and_cron_at_cond                 | 224 kB
 index_good_jobs_on_labels                                    | 16 kB
 index_good_jobs_on_locked_by_id                              | 8192 bytes
 index_good_jobs_on_priority_scheduled_at_unfinished_unlocked | 40 kB
 index_good_jobs_on_queue_name_and_scheduled_at               | 40 kB
 index_good_jobs_on_scheduled_at                              | 32 kB
 index_good_jobs_on_serialized_params_and_labels              | 48 MB
```

The new index uses 48 MB on disk. That matches the increase of the `good_jobs` total size 142 (initial) + 48 (new index) = 190 MB.

Now, it's time to search!

```
# explain (buffers on, analyze on) select count(id) from good_jobs where good_jobs_jsonb_values(serialized_params) || labels @> array['0'];
                                                         QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=61164.00..61164.01 rows=1 width=8) (actual time=13458.290..13458.290 rows=1 loops=1)
   Buffers: shared hit=9716 read=13040
   ->  Seq Scan on good_jobs  (cost=0.00..60726.74 rows=174906 width=16) (actual time=0.459..13439.193 rows=174885 loops=1)
         Filter: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{0}'::text[])
         Rows Removed by Filter: 114
         Buffers: shared hit=9716 read=13040
 Planning:
   Buffers: shared hit=4 read=1
 Planning Time: 0.337 ms
 Execution Time: 13458.325 ms
```

Disappointing! This ended up costing way more time than it should. Let's try with something more unique:

```
# explain (buffers on, analyze on) select count(id) from good_jobs where good_jobs_jsonb_values(serialized_params) || labels @> array['4b078fe3-12b2-419f-a172-78d777a55f33'];
                                                                           QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=270.06..270.07 rows=1 width=8) (actual time=0.099..0.100 rows=1 loops=1)
   Buffers: shared hit=4 read=1
   ->  Bitmap Heap Scan on good_jobs  (cost=21.67..269.90 rows=61 width=16) (actual time=0.092..0.093 rows=1 loops=1)
         Recheck Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{4b078fe3-12b2-419f-a172-78d777a55f33}'::text[])
         Heap Blocks: exact=1
         Buffers: shared hit=4 read=1
         ->  Bitmap Index Scan on index_good_jobs_on_serialized_params_and_labels  (cost=0.00..21.65 rows=61 width=0) (actual time=0.022..0.023 rows=1 loops=1)
               Index Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{4b078fe3-12b2-419f-a172-78d777a55f33}'::text[])
               Buffers: shared hit=4
 Planning:
   Buffers: shared hit=2
 Planning Time: 0.418 ms
 Execution Time: 0.142 ms
```

Ah yes, that's more like it! Let's try with something a bit different:

```
# explain (buffers on, analyze on) select count(id) from good_jobs where good_jobs_jsonb_values(serialized_params) || labels @> array['14263'];
                                                                           QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=270.06..270.07 rows=1 width=8) (actual time=0.346..0.347 rows=1 loops=1)
   Buffers: shared hit=5 read=5
   ->  Bitmap Heap Scan on good_jobs  (cost=21.67..269.90 rows=61 width=16) (actual time=0.176..0.330 rows=7 loops=1)
         Recheck Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{14263}'::text[])
         Heap Blocks: exact=6
         Buffers: shared hit=5 read=5
         ->  Bitmap Index Scan on index_good_jobs_on_serialized_params_and_labels  (cost=0.00..21.65 rows=61 width=0) (actual time=0.054..0.055 rows=7 loops=1)
               Index Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{14263}'::text[])
               Buffers: shared hit=4
 Planning:
   Buffers: shared hit=2
 Planning Time: 0.617 ms
 Execution Time: 0.424 ms

```

Let's run that same query again, but using the full-text search that already exists (query copied from using the dashboard):

```
# EXPLAIN (BUFFERS ON, ANALYZE ON) SELECT id FROM "good_jobs" WHERE ((to_tsvector('english', id::text) || to_tsvector('english', COALESCE(active_job_id::text, '')) || to_tsvector('english', serialized_params) || to_tsvector('english', COALESCE(serialized_params->>'arguments', '')) || to_tsvector('english', COALESCE(error, '')) || to_tsvector('english', COALESCE(array_to_string(labels, ' '), ''))) @@ websearch_to_tsquery('14263')) ORDER BY ts_rank((to_tsvector('english', id::text) || to_tsvector('english', COALESCE(active_job_id::text, '')) || to_tsvector('english', serialized_params) || to_tsvector('english', COALESCE(serialized_params->>'arguments', '')) || to_tsvector('english', COALESCE(error, '')) || to_tsvector('english', COALESCE(array_to_string(labels, ' '), ''))), websearch_to_tsquery('14263')) DESC, created_at desc, id desc LIMIT 25;
                                                                                  QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=146532.11..146535.03 rows=25 width=28) (actual time=5796.989..5797.044 rows=7 loops=1)
   Buffers: shared hit=12359 read=10943
   ->  Gather Merge  (cost=146532.11..146617.28 rows=730 width=28) (actual time=5783.087..5783.141 rows=7 loops=1)
         Workers Planned: 2
         Workers Launched: 2
         Buffers: shared hit=12359 read=10943
         ->  Sort  (cost=145532.09..145533.00 rows=365 width=28) (actual time=5759.972..5759.973 rows=2 loops=3)
               Sort Key: (ts_rank((((((to_tsvector('english'::regconfig, (id)::text) || to_tsvector('english'::regconfig, COALESCE((active_job_id)::text, ''::text))) || to_tsvector('english'::regconfig, serialized_params)) || to_tsvector('english'::regconfig, COALESCE((serialized_params ->> 'arguments'::text), ''::text))) || to_tsvector('english'::regconfig, COALESCE(error, ''::text))) || to_tsvector('english'::regconfig, COALESCE(array_to_string(labels, ' '::text), ''::text))), websearch_to_tsquery('14263'::text))) DESC, created_at DESC, id DESC
               Sort Method: quicksort  Memory: 25kB
               Buffers: shared hit=12359 read=10943
               Worker 0:  Sort Method: quicksort  Memory: 25kB
               Worker 1:  Sort Method: quicksort  Memory: 25kB
               ->  Parallel Seq Scan on good_jobs  (cost=0.00..145521.79 rows=365 width=28) (actual time=1948.458..5759.902 rows=2 loops=3)
                     Filter: ((((((to_tsvector('english'::regconfig, (id)::text) || to_tsvector('english'::regconfig, COALESCE((active_job_id)::text, ''::text))) || to_tsvector('english'::regconfig, serialized_params)) || to_tsvector('english'::regconfig, COALESCE((serialized_params ->> 'arguments'::text), ''::text))) || to_tsvector('english'::regconfig, COALESCE(error, ''::text))) || to_tsvector('english'::regconfig, COALESCE(array_to_string(labels, ' '::text), ''::text))) @@ websearch_to_tsquery('14263'::text))
                     Rows Removed by Filter: 58331
                     Buffers: shared hit=12241 read=10943
 Planning:
   Buffers: shared hit=84
 Planning Time: 1.814 ms
 JIT:
   Functions: 13
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 2.514 ms (Deform 1.052 ms), Inlining 0.000 ms, Optimization 1.649 ms, Emission 44.997 ms, Total 49.160 ms
 Execution Time: 5832.968 ms
```

Both queries returned 7 rows -- don't be fooled by the top-level "rows=1": the GIN index query used `count(id)`. You have to dig one level deeper to see the 7 rows.

Then, I realized that I didn't have ordering in the above query. I added the same `ORDER BY` clause to see how much it cost us (GIN index + ORDER BY, returning `*` instead of `count(id)`):

```
# EXPLAIN (BUFFERS ON, ANALYZE ON) select id from good_jobs where good_jobs_jsonb_values(serialized_params) || labels @> array['14263'] ORDER BY ts_rank((to_tsvector('english', id::text) || to_tsvector('english', COALESCE(active_job_id::text, '')) || to_tsvector('english', serialized_params) || to_tsvector('english', COALESCE(serialized_params->>'arguments', '')) || to_tsvector('english', COALESCE(error, '')) || to_tsvector('english', COALESCE(array_to_string(labels, ' '), ''))), websearch_to_tsquery('14263')) DESC, created_at desc, id desc;
                                                                                                                                                                                                                                                                         QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=380.29..380.44 rows=61 width=28) (actual time=0.817..0.819 rows=7 loops=1)
   Sort Key: (ts_rank((((((to_tsvector('english'::regconfig, (id)::text) || to_tsvector('english'::regconfig, COALESCE((active_job_id)::text, ''::text))) || to_tsvector('english'::regconfig, serialized_params)) || to_tsvector('english'::regconfig, COALESCE((serialized_params ->> 'arguments'::text), ''::text))) || to_tsvector('english'::regconfig, COALESCE(error, ''::text))) || to_tsvector('english'::regconfig, COALESCE(array_to_string(labels, ' '::text), ''::text))), websearch_to_tsquery('14263'::text))) DESC, created_at DESC, id DESC
   Sort Method: quicksort  Memory: 25kB
   Buffers: shared hit=10
   ->  Bitmap Heap Scan on good_jobs  (cost=21.67..378.48 rows=61 width=28) (actual time=0.348..0.788 rows=7 loops=1)
         Recheck Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{14263}'::text[])
         Heap Blocks: exact=6
         Buffers: shared hit=10
         ->  Bitmap Index Scan on index_good_jobs_on_serialized_params_and_labels  (cost=0.00..21.65 rows=61 width=0) (actual time=0.050..0.050 rows=7 loops=1)
               Index Cond: ((good_jobs_jsonb_values(serialized_params) || labels) @> '{14263}'::text[])
               Buffers: shared hit=4
 Planning:
   Buffers: shared hit=2
 Planning Time: 0.528 ms
 Execution Time: 0.897 ms
```

This is very interesting: use the GIN index to select rows that match, then order by full-text search. Kind of a "best of both worlds".

I then tried with a few more search terms, with more or less frequency:

```
# EXPLAIN (BUFFERS ON, ANALYZE ON) select id from good_jobs where good_jobs_jsonb_values(serialized_params) || labels @> array['America/Indiana/Indianapolis'] ORDER BY ts_rank((to_tsvector('english', id::text) || to_tsvector('english', COALESCE(active_job_id::text, '')) || to_tsvector('english', serialized_params) || to_tsvector('english', COALESCE(serialized_params->>'arguments', '')) || to_tsvector('english', COALESCE(error, '')) || to_tsvector('english', COALESCE(array_to_string(labels, ' '), ''))), websearch_to_tsquery('America/Indiana/Indianapolis')) DESC, created_at desc, id desc;
                                                                                                                                                                                                                                                                                     QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=2436.41..2437.50 rows=437 width=28) (actual time=63.555..63.577 rows=458 loops=1)
…